pyproject.toml
[tool.poetry]
name = "totopskills"
version = "0.1.0"
description = ""
authors = ["sai <saidyllapashaev@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
kafka-python = "^2.0.2"
pygame = "^2.5.2"
websocket = "^0.2.1"
websockets = "^12.0"
aiohttp = "^3.9.3"
fastapi = {extras = ["standard"], version = "^0.115.6"}
itsdangerous = "^2.2.0"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"
authlib = "^1.5.2"
python-jose = "^3.4.0"
passlib = "^1.7.4"
dramatiq = "^1.18.0"
redis = "^6.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
